pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "planfile"
version = "0.1.50"
description = "SDLC automation platform - strategic project management with CI/CD integration and automated bug-fix loops"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{name = "Tom Sapletta", email = "tom@sapletta.com"}]
keywords = ["sdlc", "strategy", "sprint", "pm", "jira", "github", "gitlab", "tickets", "agile", "cicd", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "requests>=2.31",
    "httpx>=0.27",
    "filelock>=3.0",
    "python-dotenv>=1.0",
    "PyGithub>=2.0",
]

[project.scripts]
planfile = "planfile.cli.commands:main"

[project.optional-dependencies]
github = ["PyGithub>=2.0"]
jira = ["jira>=3.5"]
gitlab = ["python-gitlab>=4.0"]
litellm = ["litellm>=1.40"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
llx = ["llx>=0.2.0"]
all = [
    "PyGithub>=2.0",
    "jira>=3.5",
    "python-gitlab>=4.0",
    "litellm>=1.40",
    "openai>=1.0",
    "anthropic>=0.25",
    "llx>=0.2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "black>=23.0",
    "isort>=5.12",
]

[project.urls]
Homepage = "https://github.com/wronai/strategy"
Repository = "https://github.com/wronai/strategy"
Documentation = "https://strategy.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["planfile*"]

[tool.setuptools.package-data]
"planfile" = ["examples/**/*.yaml", "examples/**/*.yml"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=planfile --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.code2llm]
exclude = ["htmlcov", "*.egg-info", "__pycache__", ".pytest_cache"]
